backend/requirements.txt
# FastAPI Backend Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# PyTorch and Computer Vision
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
pillow>=10.0.0
opencv-python>=4.8.0

# NLP and Text Processing
langdetect>=1.0.9
jieba>=0.42.1
transformers>=4.35.0

# Visualization and Analysis
matplotlib>=3.7.0
scikit-learn>=1.3.0

# API Models
pydantic>=2.0.0

# Optional: GPU acceleration
# torch-cuda (install separately based on CUDA version)
